backend/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
